refactor: Compose ScreenTemplate of smaller components
