+++ src/components/sign-up/screen/screen-content.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import { Box, Stack } from '@blockstack/ui';
+import { Title, Body } from '../typography';
+
+interface ScreenContentProps {
+  title: string;
+  body?: (string | JSX.Element)[];
+}
+
+export const ScreenContent: React.FC<ScreenContentProps> = ({ title, body }) => (
+  <Box px={5}>
+    <Stack spacing={2}>
+      <Title>{title}</Title>
+      <Stack spacing={[3, 4]}>
+        {body && body.length ? body.map((text, key) => <Body key={key}>{text}</Body>) : body}
+      </Stack>
+    </Stack>
+  </Box>
+);
